+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.0
+
+- Renamed `bootstrap` arguments to be more self explanatory.
+- Allow bypassing the spawn of a child process with `enableChildProcess: false`.
+- Listen to process signals from inside the worker instead of only the wrapper.
+
 ## 1.0.0
 
 - Initial release.